backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.0",
    "pyyaml>=6.0.1",
    "cryptography>=42.0.0",
    "django-environ>=0.11.0",
]
